[to be squashed]: fix spelling and test logging
