+++ src/test/java/uk/gov/pay/products/persistence/dao/PaymentDaoIT.java
@@ -25,7 +25,7 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static uk.gov.pay.products.util.RandomIdGenerator.randomInt;
 import static uk.gov.pay.products.util.RandomIdGenerator.randomUuid;
-import static uk.gov.pay.products.utils.PaymentTestHelper.createPaymentEntity;
+import static uk.gov.pay.products.utils.TestHelpers.createPaymentEntity;
 
 public class PaymentDaoIT extends DaoTestBase {
