+++ src/test/java/uk/gov/pay/products/service/PaymentDeleterTest.java
@@ -1,10 +1,18 @@
 package uk.gov.pay.products.service;
 
+import ch.qos.logback.classic.Level;
+import ch.qos.logback.classic.Logger;
+import ch.qos.logback.classic.spi.ILoggingEvent;
+import ch.qos.logback.classic.spi.LoggingEvent;
+import ch.qos.logback.core.Appender;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.mockito.ArgumentCaptor;
+import org.mockito.Captor;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
+import org.slf4j.LoggerFactory;
 import uk.gov.pay.products.config.ExpungeHistoricalDataConfig;
 import uk.gov.pay.products.fixtures.ProductEntityFixture;
 import uk.gov.pay.products.persistence.dao.PaymentDao;
@@ -14,7 +22,6 @@ import java.time.Clock;
 import java.time.Instant;
 import java.time.ZonedDateTime;
 import java.util.List;
-import java.util.stream.Collectors;
 
 import static java.time.ZoneOffset.UTC;
 import static java.util.stream.Collectors.toList;
@@ -24,7 +31,8 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
-import static uk.gov.pay.products.utils.PaymentTestHelper.createPaymentEntity;
+import static uk.gov.pay.products.utils.TestHelpers.createPaymentEntity;
+import static uk.gov.pay.products.utils.TestHelpers.verifyLog;
 
 @RunWith(MockitoJUnitRunner.class)
 public class PaymentDeleterTest {
@@ -39,8 +47,18 @@ public class PaymentDeleterTest {
 
     private final Clock clock = Clock.fixed(Instant.parse("2022-03-03T10:15:30Z"), UTC);
 
+    @Captor
+    private ArgumentCaptor<LoggingEvent> loggingEventArgumentCaptor;
+
+    @Mock
+    private Appender<ILoggingEvent> mockAppender;
+
     @Before
     public void setup() {
+        final Logger logger = (Logger) LoggerFactory.getLogger(Logger.ROOT_LOGGER_NAME);
+        logger.addAppender(mockAppender);
+        logger.setLevel(Level.INFO);
+        
         paymentDeleter = new PaymentDeleter(expungeHistoricalDataConfig, paymentDao, clock);
     }
     
@@ -48,7 +66,7 @@ public class PaymentDeleterTest {
     public void shouldDeletePayments() {
         when(expungeHistoricalDataConfig.isExpungeHistoricalDataEnabled()).thenReturn(true);
         when(expungeHistoricalDataConfig.getExpungeDataOlderThanDays()).thenReturn(1);
-        when(expungeHistoricalDataConfig.getNumberOfTransactionsToRedact()).thenReturn(3);
+        when(expungeHistoricalDataConfig.getNumberOfTransactionsToExpunge()).thenReturn(3);
 
         var payment1 = createPaymentEntity(ProductEntityFixture.aProductEntity().build(), ZonedDateTime.now(clock), 2);
         var payment2 = createPaymentEntity(ProductEntityFixture.aProductEntity().build(), ZonedDateTime.now(clock), 3);
@@ -68,5 +86,7 @@ public class PaymentDeleterTest {
         when(expungeHistoricalDataConfig.isExpungeHistoricalDataEnabled()).thenReturn(false);
         paymentDeleter.deletePayments();
         verifyNoMoreInteractions(paymentDao);
+        verifyLog(mockAppender, loggingEventArgumentCaptor, 1,
+                "Expunging of historical data is not enabled.");
     }
 }
